refactor(jira_repository): use data sources
