+++ packages/platforms_integrations/jira_repository/lib/src/jira_repository.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:jira_repository/jira_repository.dart';
-import 'package:jira_repository/src/data_sources/local_datasources/jira_integrations_storage.dart';
+import 'package:jira_repository/src/data_sources/local_data_sources/jira_integrations_storage.dart';
 import 'package:jira_repository/src/data_sources/remote_datasources/jira_platform_apis.dart';
 import 'package:jira_repository/src/mappers/mapper.dart';
 import 'package:jira_repository/src/ui/jira_platform_basic_integration_tile.dart';
